requirements.txt
cffi==1.14.0
cryptography==2.9.2
Django==2.2.13
django-rest-knox==4.1.0
djangorestframework==3.10.3
mysqlclient==2.0.1
pycparser==2.20
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
